Intercept DIAL and CALL to tel: and rewrite to cheogram
This is a proof of concept.  It catches tel: from browser, but also taps on call
button in contacts, etc, and rewrites to a Cheogram JID.  It does not query the
user's roster for gateway options, or ask the gateways to convert the URI, but
simply assumes Cheogram format and jams it in.  It also does not initiate a
call, but simply uses the default XmppUri action which will be to either add to
roster or open conversation.