Mostly rewrite dialler integration
Doing it properly this time. No Intent, no popping a second activity, just
instrument the call into the Android UI.

This turned out to be a bit easier than expected.  The ConnectionService does
indeed run inside our app process everywhere it matters, so we can reuse the
machinery from elsewhere to get a reference to XmppConnectionService.  Asking
JingleConnectionManager to send propose is enough to kick off all audio, etc.

Put in a guard that the JingleConnectionManager expects so that we won't try to
start a call if isBusy is true (that is, if another call is in progress), and
just return a failed connection BUSY in that case.

Added a flag to XmppConnectionService that we set when using the dialler
integration so it will temporarily ignore phone call state and not hang itself
up.

The app still shows a separate call notification, so you see two in-call
notifications.  You can also go into the app during the call and manually pull
up the call UI there.  If you do anything in either call UI it will work and
sync state, so that's not dangerous just perhaps odd.

Kept the connection address as a tel: for contact integration in the call
history, but use our normalized version now instead of whatever we got raw from
the system.

Add support for post-dial DTMF including 2-second pause with comma and
manual-length user-intervention pause with semicolon.  Dial with 100ms break
between tones.  Doing no break at all mostly works in practise, but not for two
identical tones in a row, this seems to be enough in my tests.

When using a post-dial string from contacts the UI shows more/different state
than it should and at least in my emulator plays a different ringback tone for
one ring before switching.  Very odd, as though it's partially ignoring our
Connection object's settings.

Audio routing has not been tested yet (for speakerphone, bluetooth, etc).