Ask user for microphone permission from dialler integration
We can't ask them for it directly.  The library I am using here will first check
if the permission is already granted.  If so onPermissionGranted will fire and
we will start the call.  If not, the Connection will be created in a dialling
state and a notification will appear saying that the app needs additional
permissions.

Once the user taps the notification they will then be asked to allow microphone
access.  If they say yes, the call will proceed.  If they deny, the call will terminate.