Check for gateway/pstn after caps has been fetched
Previously we would check for the identity at contact presence change,
but the first time the contact comes online (say, during onboarding or
first app use) the contact has never had service discovery done and so
the identity is not known.  Move the identity check to after the service
discovery has certainly been done in order to be sure it works first
time.