set in-call UI active in the connecting state
Currently, if the "CONNECTING" stage of a Jingle call takes a
significant amount of time (say, a second), then there will be a window
where the call is already accepted but the ringing UI is still showing.
This 1) is confusing; 2) allows the user to accept a call twice, which
will cause a crash.

Unfortunately, there is no native equivalent to the "CONNECTING" state.
The only sane choice here seems to be showing the in-call UI even if the
call is still connecting.