Show all available actions
Instead of a single "done" button, show all available actions in the footer,
each in a different color (generated the same as avatar backgrounds, etc) with
the action name translated if a translation is available.

The pseudo-action "close" is for when a command is already completed.  In this
case, the execute() call will see that the session is already completed and it
will be removed without actually checking the action string at all.