Default action if there are none
Spec says that actions must be specified if there are any.  In practise, at
least biboumi violates this and sends back no actions at all but is in
"executing" state meaning it expects a next stage.  So in this case we should do
something smarter than assuming we're all done.  Put up a button for the
execute/default action.