Content Addressable Media Storage
Downloads are stored at a filename that is the CID computed from a SHA256 hash
of their contents.  Where necessary (such as to stream a large file down) they
are stored at the name based on the message UUID temporarily and then renamed.

Uploads, when they must be copied into our storage, are stored in the same way.

This results in de-duplicated storage, so sending the same file back and forth
many times even in different conversations results in only one copy being
stored.

Since Android does not play nice with symlinks, but we want to be able to look
up media by multiple possible hashes, the CIDs for SHA1, SHA256, and SHA512 are
all stored in a new database table whenever a file is downloaded or uploaded,
even if the file is never copied into our storage at all (so then the db table
contains a path to, say, the image in the camera folder).

When bits-of-binary cid transfer is attempted, first check if we have that
content locally already and just use it if we already have it instead of
requesting it again.