Terrible gross hack for starting group text
This is the first feature to hardcode cheogram.com let's hope we can remove it
someday.

XEP-0033 does not have good fallback behaviour, and group text doesn't map well
to MUC.  We could still do a MUC mapping, or think of a better fallback
behaviour, or something else, but for now analysis paralysis means we stick to
the gross +1...,+1...,+1...@cheogram.com -- but that's not a protocol, it's not
something we can detect or design around or expect anyone else to ever do.  So
just check for cheogram.com because that's the only place we ever expect to see
this dirty hack.

If starting a new private group and every selected member is @cheogram.com, then
ask if we should make a group text instead.