Remove redundate oobUri, set FileParams from OOB or SIMS
The url was being stored as part of oobUri and the FileParams, which is
redundant.  Just store FileParams now, and construct it directly from the OOB
or (newly) SIMS reference if present.

This brings rudimentary support for SIMS, which saves us a file size check
roundtrip when SIMS is present.