Clear Command Buttons in Message View
We setup the buttons, but never clear them, so when a person has enough
messages in their view and one of them has a button, as they scroll
around and views get recycled the button can be attached onto any
message and not just the ones that actually specify it.

The documentation makes it a bit unclear if the adapter can be set to
null or not... so I just clear it instead. I could leave it, I suppose,
but that feels a bit weird, and the next time we go through this code
we'll just replace it with another one anyway.