Allow mutating the list of children from a derived Element
Data and Field Element subclasses expect to be able to mutate the list of
children on the original Element that they are derived from using parse.  The
way the new setChildren worked resulted in not actually mutating the original.

Change this for a bindTo that makes sure all fields reference the original (and
hence any mutation on the new object mutates the original) and also a
replaceChildren that uses clear+addAll to ensure that the existing list is
mutated rather than a new list being used.

In practise, this affected room configuration and other data forms being
properly filled out.