Store FileParams as SIMS in payloads
We no longer use the old fileParams column or serialization format for new
messages, instead storing and loading to/from a SIMS element in payloads.  If
the incoming message had a SIMS element, we preserve it as much as possible and
only add newly discovered information to it.  If the incoming message had a
different kind of attachment (such as OOB) then we synthesize a SIMS element
from the extracted data.

SIMS has no spec for storing width/heigh/duration of the final media, so use RDF
namespace from schema.org for those data.