Show fallback thumbnail if there is no image
Some thumbnails are just as good as scaling down the real image, but
some (especially blurhash) are not.  So if nothing that's "good enough" to go in
the main cache is present, try the fallback.

If there is a thumbnail we can render right now (as of this commit, just
blurhash) then render that alongside the download button.  Tapping the image
starts the download just like the button.