Support extended addressing ofrom when domain matches
Allow a message to appear as from a different jid using ofrom if the domain
matches the actual from and source domain advertises support for extended
addressing. This is safe because spoofing control is up to the source server,
and the source server could have chosen to write a different from with the same
domain if it wished.

If there is a true counterpart (set from ofrom) then display nickname like in a
MUC because this message isn't from the normal counterpart of this conversation.