Show table as list of cards when too wide
If a table has too many columns then showing them all next to each other is
awful and squishy.  So instead, when too wide based on current screen size, show
as a list of cards.  The cards are themselves a grid of two columns so that they
aren't too giantly tall, and they re-use the result field rendering for each
field in that card.

Changes on redraw such that rotating to landscape can take you to table view and
rotating back to portrait can take you back to list of cards.