Unify add contact and join channel
Instead of making the user know what kind of JID they have, just let them enter
it and we can figure that out.

Instead of forcing them to add a contact or bookmark, make it optional with a
checkbox (default checked).

Instead of assuming they want to message the contact, allow them to choose to
call instead (coupled with PSTN gateway features this is effectively a dialler).