Inverted flow happy path
When doing first-time signup, create an account on the onboarding server and
jump into register command at cheogram.  When that command is done, we expect it
will give us an OOB with xmmp:domain.tld?register which will jump us into magic
create flow for that server.  Once the new account is setup, if all we
previously had was an onboarding account we execute jidSwitch and delete the
onboarding account.

If anything goes wrong you are left with the onboarding account still tied to
cheogram and set up in the app as a "normal" account.

If the registration finishes without jumping you to a ?register URI you are left
with the onboarding account, but if you make an account then it will do the
switch after that (it tries the switch blindly, without checking if the current
backend supports it).

If you press "no thanks" you just get a spinner forever.