Allow cancel onboarding
If this is just a register not part of inversion, archive the conversation when
they say "no thanks" so they get back to their regular first start experience.

If this is part of inversion, remove the onboarding account since they aren't
going to use it and set a flag so we know to redirect to magic create instead of
back to the welcome screen.  The flag also lets us know not to offer cheogram
again right after magic create like we normally would.  Then we can clear the
flag out and move on as usual.