Allow any characters in MUC nicknames
Some things are banned by resourceprep that a user might wish to use.

If this happens, then punycode the resource so they can at least join the room.
Always send their chosen nick (not a mangled version) as <nick> in presence.
Always save their chosen nick (not a mangled version) in bookmarks (risky, since
if they try to use this bookmark from a client not supporting this it may not
let them join without changing nick).
If we get a <nick> in presence, then we respect it if it matches the punycode
decode of the resource (this is to prevent spoofing or double-naming since MUC
services won't be validating the <nick> at all right now. May relax this in the
future).
Save their actual nick (whatever we're actually showing, so <nick> if we used
that) into conversation attributes instead of assuming it will match the
resource.
Use resource for technical stuff (like PM addressing) and nick for display.

Hilight on both nick and resource so that mentions from legacy clients work.